mamo app change app

+++ mapData.php
@@ -1,9 +1,14 @@
 <?php
 
-    $showMark = $mysqli -> prepare('SELECT *,map.id AS mapId FROM map LEFT JOIN map_memo ON map.id = map_memo.map_id WHERE user_id = ? AND map_memo.id IS NULL ORDER BY mapId DESC');
+    $showMark = $mysqli -> prepare('SELECT *,map.id AS mapId FROM map LEFT JOIN map_memo ON map.id = map_memo.map_id LEFT JOIN map_bookmark ON map.id = map_bookmark.map_id WHERE user_id = ? AND map_memo.id IS NULL AND map_bookmark.id IS NULL ORDER BY mapId DESC');
     $showMark -> bind_param('i',$_SESSION['id']);
     $showMark -> execute();
     $resultMark = $showMark -> get_result();
+
+    $showMapBook = $mysqli -> prepare('SELECT *,map.id AS mapId FROM map_bookmark LEFT JOIN map ON map_bookmark.map_id = map.id WHERE map.user_id = ?');
+    $showMapBook -> bind_param('i',$_SESSION['id']);
+    $showMapBook -> execute();
+    $resultMapBook = $showMapBook -> get_result();
 
 
     if(isset($_POST['lat']) && isset($_POST['lng'])){
@@ -21,7 +26,7 @@
     }
 
     if(isset($_POST['mapDel'])){
-        $mapDel = $mysqli -> prepare('DELETE FROM map WHERE id = ?');
+        $mapDel = $mysqli -> prepare('DELETE map,map_memo,map_bookmark FROM map LEFT JOIN map_memo ON map.id = map_memo.map_id LEFT JOIN map_bookmark ON map.id = map_bookmark.map_id WHERE map.id = ?');
         $mapDel -> bind_param('i',$_POST['mapDel']);
         $mapDel -> execute();
         $mapDel -> close();
@@ -58,6 +63,19 @@
         exit;
     }
 
+    if(isset($_POST['mapVal']) && isset($_POST['bookVal'])){
+        $mapVal = $_POST['mapVal'];
+        $bookVal = $_POST['bookVal'];
+        $mapBook = $mysqli -> prepare('INSERT INTO map_bookmark(map_id,book_id) VALUES(?,?)');
+        $mapBook -> bind_param('ii',$mapVal,$bookVal);
+        $mapBook -> execute();
+        $mapBook -> close();
+        $data = array('mapVal' => $mapVal ,'bookVal' => $bookVal);
+        header('Content-type:application/json;charset=UTF-8');
+        echo json_encode($data);
+        exit;
+    }
+
     if(isset($_POST['mapId'])){
         $mapId = $_POST['mapId'];
         $mapMemoDel = $mysqli -> prepare('DELETE FROM map_memo WHERE map_id = ?');
@@ -65,6 +83,18 @@
         $mapMemoDel -> execute();
         $mapMemoDel -> close();
         $data = array('mapId' => $mapId);
+        header('Content-type:application/json;charset=UTF-8');
+        echo json_encode($data);
+        exit;
+    }
+
+    if(isset($_POST['mapBookId'])){
+        $mapBookId = $_POST['mapBookId'];
+        $mapBookDel = $mysqli -> prepare('DELETE FROM map_bookmark WHERE map_id = ?');
+        $mapBookDel -> bind_param('i',$mapBookId);
+        $mapBookDel -> execute();
+        $mapBookDel -> close();
+        $data = array('mapBookId' => $mapBookId);
         header('Content-type:application/json;charset=UTF-8');
         echo json_encode($data);
         exit;
